Implement maxSeverity tracking on log handler
