+++ asciidoctorj-core/src/main/java/org/asciidoctor/jruby/cli/MaxSeverityLogHandler.java
@@ -1,0 +1,20 @@
+package org.asciidoctor.jruby.cli;
+
+import org.asciidoctor.log.LogHandler;
+import org.asciidoctor.log.LogRecord;
+import org.asciidoctor.log.Severity;
+
+public class MaxSeverityLogHandler implements LogHandler {
+    private Severity maxSeverity = Severity.DEBUG;
+
+    @Override
+    public void log(LogRecord logRecord) {
+        if (this.maxSeverity.compareTo(logRecord.getSeverity()) < 0) {
+            this.maxSeverity = logRecord.getSeverity();
+        }
+    }
+
+    public Severity getMaxSeverity() {
+        return this.maxSeverity;
+    }
+}
